refactor-components * Move MineTavlerModal to components directory
