+++ src/components/MineTavlerModal/MineTavlerModal.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/styles/import.scss';
+@import '../../assets/styles/import.scss';
 
 .mine-tavler-modal.eds-modal__content {
     padding: 0.8rem 1.5rem 1.5rem;
