+++ src/components/MineTavlerModal/MineTavlerModal.tsx
@@ -4,14 +4,14 @@ import { Modal } from '@entur/modal'
 import { Heading3, Paragraph } from '@entur/typography'
 import { GridContainer, GridItem } from '@entur/grid'
 import { PrimaryButton, SecondaryButton } from '@entur/button'
-import { useUser } from '../../../UserProvider'
-import { useSettings } from '../../../settings/SettingsProvider'
-import { LoginModal } from '../../LoginModal/LoginModal'
-import { LoginCase } from '../../LoginModal/login-modal-types'
-import { CloseButton } from '../../CloseButton/CloseButton'
-import sikkerhetBom from '../../../assets/images/sikkerhet_bom.png'
-import retinaSikkerhetBom from '../../../assets/images/sikkerhet_bom@2x.png'
-import { getDocumentId } from '../../../utils'
+import { useUser } from '../../UserProvider'
+import { useSettings } from '../../settings/SettingsProvider'
+import { LoginModal } from '../LoginModal/LoginModal'
+import { LoginCase } from '../LoginModal/login-modal-types'
+import { CloseButton } from '../CloseButton/CloseButton'
+import sikkerhetBom from '../../assets/images/sikkerhet_bom.png'
+import retinaSikkerhetBom from '../../assets/images/sikkerhet_bom@2x.png'
+import { getDocumentId } from '../../utils'
 import './MineTavlerModal.scss'
 
 const MineTavlerModal = ({ open, onDismiss }: Props): JSX.Element | null => {
